ci: use script to install lavacli.
lavacli is not part of the used image and need to be installed.
Due to changes in the lavacli dependencies it is not possible
to connect to the master installed via apt, therefore use
the pip3 version in case of Debian stretch. If Image release
is Debian 10 or newer use apt else use pip3.

This patch fixes a regression after the kasproject
downgraded the docker base image from Debian 10 to Debian 9.

Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>