ci: Add the option to build the latest kernel
For testing it is possible to select the latest kernel
and latest xenomai-version. This is done with the variable
BUILD_OPTIONS. Trigger and project variables can overwrite
this value if necessary [1]. Isar will be patch to allow
the package version to be "latest" and independent from
the kernel version.

The build can be triggered with:

curl -X POST \
     -F token=TOKEN \
     -F "ref=REF" \
     -F "variables[BUILD_OPTIONS]=opt-linux-latest.yml:opt-xenomai-next.yml" \
     https://URL

[1]: https://docs.gitlab.com/ee/ci/variables/#priority-of-environment-variables

Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>